test: create a dedicated file to store utility functions
Move some utility functions to a new dedicated file.
It's convenient when we add test code for FinBackupConfig
and FinRestore.

In addition, move makeJobSucceeded to this file too
because this function is only used in tests.
